refactor - continue to converge on android stack jetpack compose

+++ feature/library/src/main/kotlin/ephyra/feature/library/ILibraryItem.kt
@@ -17,7 +17,7 @@ interface ILibraryItem {
     // Rich optional metadata for display badges
     val canonicalId: String?
     val coverLastModified: Long
-    val sourceStatus: Long
+    val sourceStatus: Int
 
     fun matches(constraint: String): Boolean
 }
